requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.5
astroid==2.3.3
autopep8==1.5
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
decorator==4.4.2
Django==3.0.4
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-excel-base==1.0.3
django-excel-response2==3.0.1
django-filter==2.2.0
django-six==1.0.4
djangorestframework==3.11.0
Faker==4.0.2
gunicorn==20.0.4
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
include==0.2.2
initialize==1.1.1
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
moviepy==1.0.1
mysqlclient==1.4.6
numpy==1.18.2
Pillow==7.0.0
proglog==0.1.9
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
screen==1.0.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
tqdm==4.43.0
urllib3==1.25.8
wrapt==1.11.2
xlwt==1.3.0
